+++ flamenco/job_compilers/__init__.py
@@ -17,7 +17,8 @@ def register_compiler(job_type):
 
 
 # Import subpackages to register the compilers
-from . import sleep, blender_render, blender_render_progressive, abstract_compiler, exec_command
+from . import abstract_compiler, blender_render, blender_render_progressive, blender_video_chunks, \
+    exec_command, sleep
 
 
 def compile_job(job):